+++ resources/views/components/body-parameters.blade.php
@@ -3,7 +3,8 @@
     @if($name === '[]')
         <p>
             Body: <code>{{ $parameter['type'] }}</code> {!! Parsedown::instance()->text($parameter['description'] ?? '') !!}
-            @foreach($parameter['__fields'] as $subfieldName => $subfield)
+        </p>
+        @foreach($parameter['__fields'] as $subfieldName => $subfield)
                 @if(!empty($subfield['__fields']))
                     @component('scribe::components.body-parameters', ['parameters' => [$subfieldName => $subfield], 'endpointId' => $endpointId,])
                     @endcomponent
@@ -22,11 +23,10 @@
                     </p>
                 @endif
             @endforeach
-        </p>
     @elseif(!empty($parameter['__fields']))
         <p>
         <details>
-            <summary>
+            <summary style="padding-bottom: 10px;">
                 @component('scribe::components.field-details', [
                   'name' => $parameter['name'],
                   'type' => $parameter['type'] ?? 'string',
@@ -38,7 +38,6 @@
                 ])
                 @endcomponent
             </summary>
-            <br>
             @foreach($parameter['__fields'] as $subfieldName => $subfield)
                 @if(!empty($subfield['__fields']))
                     @component('scribe::components.body-parameters', ['parameters' => [$subfieldName => $subfield], 'endpointId' => $endpointId,])