+++ src/Writing/HtmlWriter.php
@@ -62,8 +62,12 @@ class HtmlWriter
 
         // Copy assets
         $assetsFolder = __DIR__ . '/../../resources';
-        if (!is_dir($destinationFolder . "/js")) {
-            mkdir($destinationFolder."/js", 0777, true);
+        // Prune older versioned assets
+        if (is_dir($destinationFolder.'/css')) {
+            Utils::deleteDirectoryAndContents($destinationFolder.'/css');
+        }
+        if (is_dir($destinationFolder.'/js')) {
+            Utils::deleteDirectoryAndContents($destinationFolder.'/js');
         }
         Utils::copyDirectory("{$assetsFolder}/images/", "{$destinationFolder}/images");
 
@@ -94,8 +98,9 @@ class HtmlWriter
 
     protected function getMetadata(): array
     {
-        // NB:These paths are wrong for laravel type but will be set correctly by the Writer class
         $links = [];
+
+        // NB:These paths are wrong for laravel type but will be set correctly by the Writer class
         if ($this->config->get('postman.enabled', true)) {
             $links[] = '<a href="../docs/collection.json">View Postman collection</a>';
         }
@@ -120,7 +125,7 @@ class HtmlWriter
             'last_updated' => date("F j Y"),
             'auth' => $auth,
             'try_it_out' => $this->config->get('try_it_out'),
-            'links' => $links + ['<a href="http://github.com/knuckleswtf/scribe">Documentation powered by Scribe ✍</a>'],
+            'links' => array_merge($links, ['<a href="http://github.com/knuckleswtf/scribe">Documentation powered by Scribe ✍</a>']),
         ];
     }
 }