+++ src/Writing/PostmanCollectionWriter.php
@@ -303,8 +303,12 @@ class PostmanCollectionWriter
     {
         return collect($endpoint->responses)->map(function (Response $response) {
             $headers = [];
-            foreach ($response->headers as $header => $values) {
-                $headers[] = ['key' => $header, 'value' => implode('; ', $values)];
+            foreach ($response->headers as $header => $value) {
+                $headers[] = [
+                    'key' => $header,
+                    // Todo remove array support in future
+                    'value' => is_array($value) ? implode('; ', $value) : $value
+                ];
             }
 
             return [